@@ -10,21 +10,21 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_local_notifications: ^17.2.3
+  flutter_local_notifications: ^18.0.1
   cupertino_icons: ^1.0.8
-  google_maps_flutter: ^2.9.0
-  firebase_messaging: ^15.1.3
-  firebase_core: ^3.6.0
-  firebase_auth: ^5.3.1
-  firebase_storage: ^12.3.6
+  google_maps_flutter: ^2.10.0
+  firebase_messaging: ^15.1.6
+  firebase_core: ^3.8.1
+  firebase_auth: ^5.3.4
+  firebase_storage: ^12.3.7
   geofence_service: ^6.0.0+1
   geocoding: ^3.0.0
-  path_provider: ^2.1.4
+  path_provider: ^2.1.5
   provider: ^6.1.2
 #  material_design_icons_flutter: ^7.0.7296
   intl: ^0.19.0
   http: ^1.2.2
-  device_info_plus: ^10.1.2
+  device_info_plus: ^11.1.1
   image_picker: ^1.1.2
   universal_platform: ^1.1.0
   focused_menu: ^1.0.5
@@ -32,53 +32,53 @@ dependencies:
   badges: ^3.1.2
   cached_network_image: ^3.4.1
   animated_splash_screen: ^1.3.0
-  font_awesome_flutter: ^10.7.0
+  font_awesome_flutter: ^10.8.0
   page_transition: ^2.1.0
   responsive_builder: ^0.7.1
   animations: ^2.0.11
   flex_color_scheme: ^7.3.1
   pin_code_fields: ^8.0.1
-  geolocator: ^13.0.1
+  geolocator: ^13.0.2
   maps_toolkit: ^3.0.0
-  flutter_riverpod: ^2.5.1
+#  flutter_riverpod: ^2.5.1
   pretty_json: ^2.0.0
   dart_geohash: ^2.1.0
   proximity_hash: ^2.1.0
   fluttertoast: ^8.2.8
-  equatable: ^2.0.5
+  equatable: ^2.0.7
   permission_handler: ^11.3.1
   geojson: ^1.0.0
   native_device_orientation: ^2.0.3
-  image: ^4.2.0
-  firebase_dynamic_links: ^6.0.8
+  image: ^4.3.0
+  firebase_dynamic_links: ^6.0.11
   get: ^4.6.6
   flutter_easyloading: ^3.0.5
-  firebase_ui_auth: ^1.15.0
-  sms_autofill: ^2.4.0
+  firebase_ui_auth: ^1.16.0
+  sms_autofill: ^2.4.1
   numeric_keyboard: ^1.1.0
   flutter_analog_clock: ^1.0.3
   custom_map_markers: ^0.0.2+1
   google_maps_cluster_manager: ^3.1.0
-  firebase_analytics: ^11.3.3
-  url_launcher: ^6.3.0
-  flutter_dotenv: ^5.1.0
+  firebase_analytics: ^11.3.6
+  url_launcher: ^6.3.1
+  flutter_dotenv: ^5.2.1
   dots_indicator: ^3.0.0
   google_fonts: ^6.2.1
-  shared_preferences: ^2.3.2
-  package_info_plus: ^8.0.2
+  shared_preferences: ^2.3.3
+  package_info_plus: ^8.1.1
   archive: ^3.6.1
   universal_io: ^2.2.2
   pinput: ^5.0.0
   intl_phone_field: ^3.2.0
-  file_picker: ^8.1.2
+  file_picker: ^8.1.4
   scrollable_positioned_list: ^0.3.8
 #  mobile_scanner: ^6.0.2
-  get_it: ^8.0.1
+  get_it: ^8.0.2
   getwidget: ^4.0.0
   json_annotation: ^4.9.0
   mongo_dart: ^0.10.3
   sembast_web: ^2.4.0+4
-  sembast: ^3.7.5+2
+  sembast: ^3.8.0+1
   uuid: any
   theme_mode_handler: ^3.0.0
   flutter_image_compress: ^2.3.0
@@ -90,6 +90,8 @@ dependencies:
   simple_barcode_scanner: ^0.2.5
   pretty_print_json: ^1.0.1
   widgets_to_image: ^1.0.0
+  get_thumbnail_video: ^0.7.3
+
 
 dev_dependencies:
   flutter_test:
