@@ -1,5 +1,5 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/foundation.dart';
+import 'package:firebase_storage/firebase_storage.dart';
 import 'package:get_it/get_it.dart';
 import 'package:http/http.dart' as http;
 import 'package:kasie_transie_library/bloc/cloud_storage_bloc.dart';
@@ -16,6 +16,7 @@ import '../utils/error_handler.dart';
 import '../utils/functions.dart';
 import '../utils/prefs.dart';
 import '../utils/route_distance_calculator.dart';
+import '../widgets/scanners/qr_code_generation.dart';
 import 'app_auth.dart';
 import 'cache_manager.dart';
 import 'data_api_dog.dart';
@@ -26,17 +27,21 @@ class RegisterServices {
   static String dbPath = 'kasie.db';
   static DatabaseFactory dbFactoryWeb = databaseFactoryWeb;
 
-  static Future<void> register() async {
+  static Future<String> register( {required FirebaseStorage firebaseStorage}) async {
     pp('\n\n$mm  ... initialize service singletons with GetIt .... 🍎🍎🍎');
-
+    pp('$mm .... QRGeneration: 🦠qrgGeneration initialized');
     final http.Client client = http.Client();
+    pp('$mm .... http.Client: 🦠client initialized');
     final AppAuth appAuth = AppAuth(FirebaseAuth.instance);
     pp('$mm .... AppAuth: 🦠auth initialized');
-
+    final DeviceLocationBloc deviceLocationBloc = DeviceLocationBloc();
+    pp('$mm .... DeviceLocationBloc: 🦠deviceLocationBloc initialized');
     final CacheManager cacheManager = CacheManager();
+    pp('$mm .... CacheManager: 🦠cacheManager initialized');
     final Prefs prefs = Prefs(await SharedPreferences.getInstance());
+    pp('$mm .... Prefs: 🦠prefs initialized');
     final ErrorHandler errorHandler = ErrorHandler(DeviceLocationBloc(), prefs);
-
+    pp('$mm .... ErrorHandler: 🦠errorHandler initialized');
     final SemCache semCache = SemCache();
     pp('$mm .... SemCache: 🦠cache initialized');
     final ZipHandler zipHandler = ZipHandler(appAuth, semCache);
@@ -48,58 +53,65 @@ class RegisterServices {
         ListApiDog(client, appAuth, prefs, errorHandler, zipHandler, semCache);
     pp('$mm .... ListApiDog: 🦠listApiDog initialized');
     //
-    pp('$mm ..... 🦠🦠🦠🦠🦠registerLazySingletons ...');
+    CloudStorageBloc csb = CloudStorageBloc(dataApiDog: dataApi,
+        prefs: prefs, firebaseStorage: firebaseStorage,
+      locationBloc: deviceLocationBloc, );
+    pp('$mm .... CloudStorageBloc: 🦠csb initialized');
+    pp('\n\n$mm ..... 🦠🦠🦠🦠🦠registerLazySingletons ...');
+    final QRGenerationService qrgGenerationService = QRGenerationService(csb);
 
-    GetIt.instance.registerLazySingleton<KasieThemeManager>(() => KasieThemeManager(prefs));
+    final GetIt instance = GetIt.instance;
+
+    instance.registerLazySingleton<Prefs>(() => prefs);
+    pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... Prefs');
+
+    instance.registerLazySingleton<CloudStorageBloc>(() => csb);
+    pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... CloudStorageBloc');
+
+    instance.registerLazySingleton<KasieThemeManager>(() => KasieThemeManager(prefs));
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... KasieThemeManager');
 
-    GetIt.instance.registerLazySingleton<RouteUpdateListener>(() => RouteUpdateListener());
+    instance.registerLazySingleton<RouteUpdateListener>(() => RouteUpdateListener());
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... RouteUpdateListener');
 
-    GetIt.instance.registerLazySingleton<DeviceLocationBloc>(() => DeviceLocationBloc());
+    instance.registerLazySingleton<DeviceLocationBloc>(() => deviceLocationBloc);
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... DeviceLocationBloc');
 
-    GetIt.instance.registerLazySingleton<SemCache>(() => semCache);
+    instance.registerLazySingleton<SemCache>(() => semCache);
 
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... SemCache');
 
-    GetIt.instance.registerLazySingleton<ZipHandler>(() => zipHandler);
+    instance.registerLazySingleton<ZipHandler>(() => zipHandler);
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... ZipHandler');
 
-    GetIt.instance.registerLazySingleton<RouteDistanceCalculator>(
+    instance.registerLazySingleton<RouteDistanceCalculator>(
         () => RouteDistanceCalculator(prefs, listApi, dataApi));
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... RouteDistanceCalculator');
 
-    GetIt.instance.registerLazySingleton<CloudStorageBloc>(
-        () => CloudStorageBloc(dataApi, prefs));
-    pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... CloudStorageBloc');
-
-    GetIt.instance.registerLazySingleton<TheGreatGeofencer>(
+    instance.registerLazySingleton<TheGreatGeofencer>(
         () => TheGreatGeofencer(dataApi, listApi, prefs));
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... TheGreatGeofencer');
 
-    GetIt.instance.registerLazySingleton<KasieThemeManager>(() => KasieThemeManager(prefs));
-    pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... ThemeBloc');
-
-    GetIt.instance.registerLazySingleton<ListApiDog>(() => listApi);
+    instance.registerLazySingleton<ListApiDog>(() => listApi);
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... ListApiDog');
 
-    GetIt.instance.registerLazySingleton<DataApiDog>(() => dataApi);
+    instance.registerLazySingleton<DataApiDog>(() => dataApi);
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... DataApiDog');
 
-    GetIt.instance.registerLazySingleton<CacheManager>(() => cacheManager);
+    instance.registerLazySingleton<CacheManager>(() => cacheManager);
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... CacheManager');
 
-    GetIt.instance.registerLazySingleton<Prefs>(() => prefs);
-    pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... Prefs');
-
-    GetIt.instance.registerLazySingleton<AppAuth>(() => appAuth);
+    instance.registerLazySingleton<AppAuth>(() => appAuth);
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... AppAuth');
 
-    GetIt.instance.registerLazySingleton<ErrorHandler>(() => errorHandler);
+    instance.registerLazySingleton<ErrorHandler>(() => errorHandler);
     pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... ErrorHandler');
 
+    instance.registerLazySingleton<QRGenerationService>(() => qrgGenerationService);
+    pp('$mm 🦠🦠🦠🦠🦠registerLazySingletons ... QRGenerationService');
 
-    pp('\n\n$mm   🍎🍎🍎 13 Service singletons registered! .... 🍎🍎🍎\n');
+
+    pp('\n\n$mm  returning message form RegisterService  🍎🍎🍎\n\n');
+    return '\n🍎🍎🍎 RegisterServices: 15 Service singletons registered!';
   }
 }
