@@ -578,18 +578,21 @@ Vehicle _$VehicleFromJson(Map<String, dynamic> json) => Vehicle(
       ownerName: json['ownerName'] as String?,
       ownerId: json['ownerId'] as String?,
       created: json['created'] as String?,
+      qrCodeBytes: json['qrCodeBytes'] as String?,
       dateInstalled: json['dateInstalled'] as String?,
       vehicleReg: json['vehicleReg'] as String?,
       make: json['make'] as String?,
       model: json['model'] as String?,
       year: json['year'] as String?,
       qrCodeUrl: json['qrCodeUrl'] as String?,
-      ownerCellphone: json['ownerCellphone'] as String?,
+      cellphone: json['cellphone'] as String?,
+      active: (json['active'] as num?)?.toInt(),
       passengerCapacity: (json['passengerCapacity'] as num?)?.toInt(),
       associationId: json['associationId'] as String?,
       photos: (json['photos'] as List<dynamic>?)
           ?.map((e) => VehiclePhoto.fromJson(e as Map<String, dynamic>))
           .toList(),
+      bucketFileName: json['bucketFileName'] as String?,
       videos: (json['videos'] as List<dynamic>?)
           ?.map((e) => VehicleVideo.fromJson(e as Map<String, dynamic>))
           .toList(),
@@ -601,7 +604,7 @@ Map<String, dynamic> _$VehicleToJson(Vehicle instance) => <String, dynamic>{
       'countryId': instance.countryId,
       'ownerName': instance.ownerName,
       'ownerId': instance.ownerId,
-      'ownerCellphone': instance.ownerCellphone,
+      'cellphone': instance.cellphone,
       'created': instance.created,
       'dateInstalled': instance.dateInstalled,
       'vehicleReg': instance.vehicleReg,
@@ -609,7 +612,10 @@ Map<String, dynamic> _$VehicleToJson(Vehicle instance) => <String, dynamic>{
       'model': instance.model,
       'year': instance.year,
       'qrCodeUrl': instance.qrCodeUrl,
+      'bucketFileName': instance.bucketFileName,
       'passengerCapacity': instance.passengerCapacity,
+      'active': instance.active,
+      'qrCodeBytes': instance.qrCodeBytes,
       'associationId': instance.associationId,
       'associationName': instance.associationName,
       'photos': instance.photos?.map((e) => e.toJson()).toList(),
@@ -776,8 +782,10 @@ User _$UserFromJson(Map<String, dynamic> json) => User(
       associationName: json['associationName'] as String?,
       fcmToken: json['fcmToken'] as String?,
       password: json['password'] as String?,
+      bucketFileName: json['bucketFileName'] as String?,
       email: json['email'] as String?,
       cellphone: json['cellphone'] as String?,
+      qrCodeBytes: json['qrCodeBytes'] as String?,
       profileThumbnail: json['profileThumbnail'] as String?,
       created: json['created'] as String?,
       qrCodeUrl: json['qrCodeUrl'] as String?,
@@ -796,6 +804,8 @@ Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{
       'fcmToken': instance.fcmToken,
       'password': instance.password,
       'email': instance.email,
+      'bucketFileName': instance.bucketFileName,
+      'qrCodeBytes': instance.qrCodeBytes,
       'cellphone': instance.cellphone,
       'profileThumbnail': instance.profileThumbnail,
       'profileUrl': instance.profileUrl,
